requirements.txt
customtkinter==5.1.2
Pillow==9.4.0
pynput==1.7.6
